finished csound evoke

+++ rpr/exec_csound-ATS.py
@@ -4,6 +4,9 @@ import os, sys
 import subprocess
 import time
 import re
+import concurrent.futures
+
+REMOVE_FILEs = True
 
 def extract_score_data(string):
 	pattern = r'/\*([\s\S]*?)\*/'
@@ -14,7 +17,20 @@ def extract_score_data(string):
 	else:
 		return None
 
+def run_atsa(command):
+	try:
+		process = subprocess.Popen(command, stdout=subprocess.PIPE, stderr=subprocess.PIPE)
+		stdout, stderr = process.communicate()
+		if process.returncode != 0:
+			print(f"An error occurred: {stderr.decode('utf-8')}")
+		else:
+			print(f'Csound processing {input_file} completed successfully.')
+	except subprocess.CalledProcessError as e:
+		return f"Error running command: {command}\n{e.stderr.strip()}"
+	
+
 cs = ctcsound.Csound()
+cs.createMessageBuffer(False)
 
 input_file_wav = sys.argv[1]
 input_file_orc = sys.argv[2]
@@ -29,6 +45,7 @@ sample_rate = sox.file_info.sample_rate(input_file_wav)
 #output_tempdir = os.path.dirname(file)
 output_tempdir = '/Users/j/Documents/PROJECTs/_temp'
 
+log_file = output_file_wav + '.log'
 
 
 with open(input_file_orc, 'r') as f:
@@ -41,6 +58,7 @@ ats_files = []
 mono_files = []
 
 channels = sox.file_info.channels(input_file_wav)
+
 for i in range(1, channels+1):
 	tfm = sox.Transformer()
 	output_file = os.path.join(output_tempdir, basename + f'-{i}ch.wav')
@@ -49,17 +67,23 @@ for i in range(1, channels+1):
 	tfm.build(input_filepath=input_file_wav, output_filepath=output_file)
 	mono_files.append(output_file)
 
+csound_commands = []
 for f in mono_files:
 	input_file = f
 	output_file = os.path.join(output_tempdir, f'{os.path.splitext(os.path.basename(input_file))[0]}.ats')
-	csound_command = ['atsa', input_file, output_file]
-	process = subprocess.Popen(csound_command, stdout=subprocess.PIPE, stderr=subprocess.PIPE)
-	stdout, stderr = process.communicate()
-	if process.returncode != 0:
-		print(f"An error occurred: {stderr.decode('utf-8')}")
-	else:
-		print(f'Csound processing {input_file} completed successfully.')
+	csound_commands.append(['atsa', input_file, output_file])
 
+with concurrent.futures.ThreadPoolExecutor() as executor:
+	# Using a list comprehension to start all commands concurrently.
+	# This returns a list of futures.
+	futures = [executor.submit(run_atsa, command) for command in csound_commands]
+
+	# Wait for all the commands to finish and retrieve the results.
+	results = [future.result() for future in concurrent.futures.as_completed(futures)]
+	# Print the results
+	for i, result in enumerate(results, start=1):
+		print(f"Command {i} output: {result}")
+	
 for f in os.listdir(output_tempdir):
 	if f.endswith('.ats') and basename in f:
 		ats_files.append('"' + os.path.join(output_tempdir, f) + '"')
@@ -68,7 +92,7 @@ score = []
 if sco_python_code is not None:
 	try:
 		for i, ats_file in enumerate(ats_files):
-			ch = int(re.search(r'(\d+)ch', ats_file)[1])
+			ch = int(re.findall(r'(\d+)ch', ats_file)[-1])
 			exec(sco_python_code)
 	except Exception as e:
 		print("Error executing the code:", str(e))
@@ -89,8 +113,17 @@ cs.compileOrc(orc_code)
 cs.readScore(score)
 
 cs.start()
-cs.perform()
+
+with open(log_file, 'a') as f:
+	while cs.performKsmps() == 0:
+		string = cs.firstMessage()
+		# Set the custom performance callback
+		if string:
+			f.write(string)
+		cs.popFirstMessage()
+
 cs.cleanup()
+cs.destroyMessageBuffer()
 del cs
 
 time.sleep(1/8)
@@ -99,15 +132,15 @@ with open(output_file_wav + '--finished', 'w') as f:
 	f.write('I EXIST')
 
 try:
-	# Remove the file
-	
-	os.remove(input_file_wav)
 	with open(input_file_orc, 'w') as f:
 		f.write(orc_code)
-	for f in mono_files:
-		os.remove(f)
-	for f in ats_files:
-		os.remove(f.replace('"', ''))
+	if REMOVE_FILEs:
+		# Remove the file
+		os.remove(input_file_wav)
+		for f in mono_files:
+			os.remove(f)
+		for f in ats_files:
+			os.remove(f.replace('"', ''))
 
 	print("File removed successfully.")
 except FileNotFoundError:
